% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selLasso.R
\name{selLasso}
\alias{selLasso}
\title{Variable Selection with Lasso}
\usage{
selLasso(X, Y, target = NULL, nfolds = 10, ...)
}
\arguments{
\item{X}{numeric design matrix (excluding the intercept), where columns correspond to variables, and rows to observations.}

\item{Y}{numeric response vector.}

\item{target}{maximum number of variables to be selected.}

\item{nfolds}{number of folds to be used in the cross-validation.}

\item{...}{further arguments to be passed to \code{\link{hdi::lasso.cv}}.}
}
\value{
\code{selLasso} returns a numeric vector containing the indices of the selected variables.
}
\description{
This function uses the Lasso to select variables in the framework of linear regression.
}
\details{
If target is not null, the function uses the Lasso with suitable calibration of the \code{lambda} parameter to select at most \code{target} variables.
Otherwise, it uses the Lasso with 10-fold cross-validation.
}
\examples{
# generate linear regression data with 20 variables and 10 observations
res <- simData(m1=2, m=20, n=10, rho=0.5, type="toeplitz", SNR=5, seed=42)
X <- res$X # design matrix
Y <- res$Y # response vector
active <- res$active # indices of active variables

# choose target as twice the number of active variables
target <- 2*length(active)

# Lasso selecting at most target variables
selLasso(X, Y, target)

# Lasso with 10-fold cross-validation
selLasso(X, Y)
}
\author{
Anna Vesely.
}
