% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selOracle.R
\name{selOracle}
\alias{selOracle}
\title{Variable Selection with Oracle}
\usage{
selOracle(X, Y, target, toSel)
}
\arguments{
\item{X}{numeric design matrix (excluding the intercept), where columns correspond to variables, and rows to observations.}

\item{Y}{numeric response vector.}

\item{target}{maximum number of variables to be selected.}

\item{toSel}{numeric vector containing the indices of the variables to be always selected.}
}
\value{
\code{selOracle} returns a numeric vector containing the indices of the selected variables.
Variables in \code{toSel} are always included.
}
\description{
This function selects a given number of variables, always including those in a pre-specified set.
}
\details{
An error message is returned if \code{target} is smaller than the length of \code{toSel}.
}
\examples{
# generate linear regression data with 20 variables and 10 observations
res <- simData(m1=2, m=20, n=10, rho=0.5, type="toeplitz", SNR=5, seed=42)
X <- res$X # design matrix
Y <- res$Y # response vector
active <- res$active # indices of active variables

# choose target as twice the number of active variables
target <- 2*length(active)

# selection of at most target variables, including the active ones
selOracle(X, Y, target, active)
}
\author{
Anna Vesely.
}
