% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simData2.R
\name{simData2}
\alias{simData2}
\title{Simulating Linear Regression Data (2)}
\usage{
simData2(prop, m, n, rho = 0, alpha = 0.05, pw = 0.9, concordant = TRUE, seed = NULL)
}
\arguments{
\item{prop}{proportion of active variables.}

\item{m}{number of variables.}

\item{n}{numer of observations.}

\item{rho}{level of equicorrelation between pairs of variables.}

\item{alpha}{significance level.}

\item{pw}{power of the t test.}

\item{concordant}{logical, \code{TRUE} for positive coefficients,
\code{FALSE} for coefficients with alternating signs.}

\item{seed}{seed.}
}
\value{
\code{simData} returns a list containing the design matrix \code{X} (excluding the intercept),
the response vector \code{Y}, and the index vector of active variables \code{active}.
}
\description{
This function simulates a design matrix and a response vector.
}
\details{
The design matrix \code{X} contains \code{n} independent observations from a
MVN with mean 0 and equi-correlation \code{rho}.

A proportion \code{prop} of the coefficients are non-null.
The value of non-null coefficients is such that the one-sample t test with significance level \code{alpha},
using half the sample size, has power equal to \code{pw}. Then the response variable \code{Y}
is equal to \code{X \%\% beta} plus a standard normal error.
}
\examples{
# generate linear regression data with 30 variables and 20 observations
res <- simData2(prop=0.05, m=30, n=20, seed=42)
X <- res$X # design matrix
Y <- res$Y # response vector
active <- res$active # indices of active variables

# choose target as twice the number of active variables
target <- 2*length(active)

# matrix of standardized scores for all variables (columns) and random sign flips (rows)
# using the exact method with Lasso selection
G <- splitFlip(X, Y, target=target, exact=TRUE, seed=42)

# maxT algorithm
maxT(G)
}
\author{
Anna Vesely.
}
