% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitFlip_sign.R
\name{splitFlip_sign}
\alias{splitFlip_sign}
\title{Resampling-Based Multisplit (signed)}
\usage{
splitFlip_sign(X, Y, Q = 50, B = 200, target = NULL, varSel = selLasso, varSelArgs = NULL, exact = FALSE, maxRepeat = 20, seed = NULL, sign_type = 0)
}
\arguments{
\item{X}{numeric design matrix (including the intercept), where columns correspond to variables, and rows to observations.}

\item{Y}{numeric response vector.}

\item{Q}{numer of data splits.}

\item{B}{number of sign flips.}

\item{target}{maximum number of variables to be selected.}

\item{varSel}{a function to perform variable selection. It must have at least three arguments:
\code{X} (design matrix), \code{Y} (response vector) and \code{target} (maximum number of selected variables).
Additional arguments are passed through \code{varSelArgs}.
Return value is a numeric vector containing the indices of the selected variables.}

\item{varSelArgs}{named list of further arguments for \code{varSel}.}

\item{exact}{logical, \code{TRUE} for the exact method, \code{FALSE} for the approximate method.}

\item{maxRepeat}{maximum number of split trials.}

\item{seed}{seed.
#' @param sign_type type of method; 0 for unsigned, 1 for the mode of the coefficient's signs, 2 for the individual signs.}
}
\value{
\code{splitFlip} returns a numeric matrix of standardized scores, where columns correspond to variables,
and rows to \code{B} random sign flips. The first flip is the identity.
}
\description{
This function computes resampling-based standardized scores for high-dimensional linear regression.
}
\details{
The data are iteratively split into two subsets of equal size for \code{Q} times.
For each split, the first subset is used to perform variable selection,
while the second is used to compute the effective scores for
each variable and \code{B} random sign flips (including the identity).
If a variable is not selected, its score is set to zero. For each variable and each sign flip, the standardized score is defined as (an approximation of)
the sum of the effective scores over the \code{Q} splits, divided by its variance.

If too many variables are selected in a split (more than half the sample size),
a warning is returned and the data is randomly split again.
After \code{maxRepeat} trials where too many variables are selected, the function returns an error message.
}
\examples{
# generate linear regression data with 20 variables and 10 observations
res <- simData(m1=2, m=20, n=10, rho=0.5, type="toeplitz", SNR=5, seed=42)
X <- res$X # design matrix
Y <- res$Y # response vector
active <- res$active # indices of active variables

# choose target as twice the number of active variables
target <- 2*length(active)

# standardized scores using the approximate method with Lasso selection of target variables
G1 <- splitFlip_sign(X, Y, target=target, seed=42, sign_type=0)

# maxT algorithm
maxT(abs(G1), alpha=0.1)

# standardized scores using the exact method with oracle selection of target variables
G2 <- splitFlip_sign(X, Y, target=target, varSel=selOracle, varSelArgs=list(toSel=active), seed=42, sign_type=0)

# maxT algorithm
maxT(abs(G2), alpha=0.1)
}
\author{
Anna Vesely.
}
