% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targetLasso.R
\name{targetLasso}
\alias{targetLasso}
\title{Variable Selection with Lasso}
\usage{
targetLasso(X, Y, target = NULL)
}
\arguments{
\item{X}{numeric design matrix (excluding the intercept), where columns correspond to variables, and rows to observations.}

\item{Y}{numeric response vector.}

\item{target}{number of variables to be selected at most. If null, it is set to \code{nrow(X)}.}
}
\value{
\code{targetLasso} returns a numeric vector containing the indices of the selected variables.
}
\description{
This function uses the Lasso to select a given number of variables in the framework of linear regression.
}
\details{
The \code{lambda} parameter is chosen as the smallest parameter that selects at most \code{target} variables.

Notice that the number of selected variables will never exceed the number of observations.
}
\examples{
set.seed(42)
n <- 5 # observations
m <- 10 # variables
active <- c(1,2)
beta <- rep(0,m)
beta[active] <- 5
X <- matrix(rnorm(m*n), ncol=m)
Y <- rnorm(n=n, mean=X \%*\% beta)
target <- 4
targetLasso(X, Y, target)
}
\author{
Anna Vesely.
}
