% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targetOracle.R
\name{targetOracle}
\alias{targetOracle}
\title{Variable Selection with Oracle}
\usage{
targetOracle(X, Y, target = NULL, m, active)
}
\arguments{
\item{X}{numeric design matrix (excluding the intercept), where columns correspond to variables, and rows to observations.}

\item{Y}{numeric response vector.}

\item{target}{number of variables to be selected at most. If null, it is set to \code{nrow(X)}.}

\item{m}{total number of variables.}

\item{active}{numeric vector containing the indices of the variables to be always selected.}
}
\value{
\code{targetOracle} returns a numeric vector containing the indices of the selected variables.
Variables in \code{active} are always included.
}
\description{
This function selects a given number of variables, always including those in a pre-specified set.
}
\details{
An error message is returned if \code{target} is smaller than the length of \code{active}.
}
\examples{
set.seed(42)
n <- 5 # observations
m <- 10 # variables
active <- c(1,2)
beta <- rep(0,m)
beta[active] <- 5
X <- matrix(rnorm(m*n), ncol=m)
Y <- rnorm(n=n, mean=X \%*\% beta)
target <- 4
targetOracle(X, Y, target, m, active)
}
\author{
Anna Vesely.
}
