% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxT.R
\name{maxT}
\alias{maxT}
\title{Single-Step maxT}
\usage{
maxT(G, alpha = 0.05)
}
\arguments{
\item{G}{numeric matrix of statistics, where columns correspond to variables, and rows to data transformations.
The first transformation is the identity.}

\item{alpha}{significance level.}
}
\value{
\code{maxT} returns a list containing the p-values for each variable, and the indices of rejected variables.
}
\description{
This function applies the single-step maxT algorithm.
}
\examples{
# generate linear regression data with 20 variables and 10 observations
res <- simData(prop=0.1, m=20, n=10, rho=0.5, type="toeplitz", seed=42)
X <- res$X # design matrix
Y <- res$Y # response vector
active <- res$active # indices of active variables

# choose target as twice the number of active variables
target <- 2*length(active)

# matrix of standardized scores for all variables (columns) and random sign flips (rows)
# using the approximate method with oracle selection
G <- splitFlip(X, Y, target=target, varSel=targetOracle, varSelArgs=list(toSel=active), seed=42)

# maxT algorithm
maxT(G, alpha=0.1)
}
\author{
Anna Vesely.
}
