% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simData.R
\name{simData}
\alias{simData}
\title{Simulating Linear Regression Data}
\usage{
simData(prop, m, n, rho = 0, alpha = 0.05, pw = 0.8, concordant = TRUE, seed = NULL)
}
\arguments{
\item{prop}{proportion of active variables.}

\item{m}{number of variables.}

\item{n}{numer of observations.}

\item{rho}{level of equicorrelation between pairs of variables.}

\item{alpha}{significance level.}

\item{pw}{power of the t test.}

\item{concordant}{logical, \code{TRUE} for positive coefficients,
\code{FALSE} for coefficients with alternating signs.}

\item{seed}{seed.}
}
\value{
\code{simData} returns a list containing the covariate matrix \code{X} (excluding the intercept),
the response vector \code{Y}, and the index vector of active variables \code{active.}
}
\description{
This function simulates a design matrix and a response vector.
}
\details{
The covariate matrix \code{X} contains \code{n} independent observations from a
MVN with mean 0 and equi-correlation \code{rho}.

The intercept and a proportion \code{prop} of the entries in the coefficient vector
are non-null, with signal such that the one-sample t test with significance level \code{alpha} has power equal to \code{pw}.
}
\examples{
# generate linear regression data with 20 variables and 10 observations
res <- simData(prop=0.1, m=20, n=10, seed=42)

# choose target as twice the number of active variables
target <- 2*length(res$active)

# matrix of standardized scores for all variables (columns) and random sign flips (rows)
# using the exact method with 10 data splits and Lasso selection
G1 <- splitFlip(res$X, res$Y, target=target, B=6, exact=TRUE)
round(G1,2)

# matrix of standardized scores for all variables (columns) and random sign flips (rows)
# using the approximate method with 5 data splits and oracle selection
G2 <- splitFlip(res$X, res$Y, target=target, Q=5, B=6,
               varSel=targetOracle, varSelArgs=list(m=ncol(res$X), active=res$active))
round(G2,2)
}
\author{
Anna Vesely.
}
