% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simData.R
\name{simData}
\alias{simData}
\title{Simulating Linear Regression Data}
\usage{
simData(m1, m, n, rho = 0, type = "equicorr", incrBeta = FALSE, SNR = 1, seed = NULL)
}
\arguments{
\item{m1}{number of active variables.}

\item{m}{total number of variables.}

\item{n}{number of observations.}

\item{rho}{correlation parameter.}

\item{type}{type of covariance matrix among \code{equicorr} and \code{toeplitz}.}

\item{incrBeta}{logical, \code{TRUE} for increasing active coefficients (1,2,3,...),
\code{FALSE} for active coefficients all equal to 1.}

\item{SNR}{signal-to-noise-ratio (ratio between the variances of \code{X beta} and the error term).}

\item{seed}{seed.}
}
\value{
\code{simData} returns a list containing the design matrix \code{X} (not including the intercept),
the response vector \code{Y}, and the index vector of active variables \code{active}.
}
\description{
This function simulates a design matrix and a response vector.
}
\details{
The design matrix \code{X} contains \code{n} independent observations from a
MVN with mean 0 and covariance matrix \code{Sigma}. The term \code{Sigma(ij)} is given by \code{type}:
\itemize{
\item equicorrelation: 1 if \code{i=j}, and \code{rho} otherwise
\item Toeplitz: \code{rho^|i-j|}
}

A number \code{m1} of the coefficients are non-null, with values depending on \code{incrBeta}.
Then the response variable \code{Y} is equal to \code{X beta} plus an error term.
The standard deviation of this error term is such that the signal-to-noise ratio is \code{SNR}.
}
\examples{
# generate linear regression data with 20 variables and 10 observations
res <- simData(m1=2, m=20, n=10, rho=0.5, type="toeplitz", SNR=5, seed=42)
X <- res$X # design matrix
Y <- res$Y # response vector
active <- res$active # indices of active variables

# choose target as twice the number of active variables
target <- 2*length(active)

# standardized scores using the approximate method with Lasso selection of target variables
G <- splitFlip(X, Y, target=target, seed=42)

# maxT algorithm
maxT(G, alpha=0.1)
}
\author{
Anna Vesely.
}
