% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targetLasso.R
\name{targetLasso}
\alias{targetLasso}
\title{Variable Selection with Lasso}
\usage{
targetLasso(X, Y, target = NULL)
}
\arguments{
\item{X}{numeric design matrix (excluding the intercept), where columns correspond to variables, and rows to observations.}

\item{Y}{numeric response vector.}

\item{target}{maximum number of variables, i.e. columns of \code{X}, to be selected. If null, it is set to \code{nrow(X)}.}
}
\value{
\code{targetLasso} returns a numeric vector containing the indices of the selected variables.
}
\description{
This function uses the Lasso to select a given number of variables in the framework of linear regression.
}
\details{
The \code{lambda} parameter is chosen as the smallest parameter that selects at most \code{target} variables.

Notice that the number of selected variables cannot exceed the number of observations.
}
\examples{
# generate linear regression data with 20 variables and 10 observations
res <- simData(prop=0.1, m=20, n=10, rho=0.5, type="toeplitz", seed=42)
X <- res$X # design matrix
Y <- res$Y # response vector
active <- res$active # indices of active variables

# choose target as twice the number of active variables
target <- 2*length(active)

# selection of at most target variables using the Lasso
targetLasso(X, Y, target)
}
\author{
Anna Vesely.
}
