% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targetOracle.R
\name{targetOracle}
\alias{targetOracle}
\title{Variable Selection with Oracle}
\usage{
targetOracle(X, Y, target = NULL, m, active, seed = NULL)
}
\arguments{
\item{X}{numeric design matrix (including the intercept), where columns correspond to variables, and rows to observations.}

\item{Y}{numeric response vector.}

\item{target}{maximum number of variables to be selected. If null, it is set to \code{nrow(X)}.}

\item{m}{total number of variables.}

\item{active}{numeric vector containing the indices of the variables to be always selected.}
}
\value{
\code{targetOracle} returns a numeric vector containing the indices of the selected variables.
Variables in \code{active} are always included.
}
\description{
This function selects a given number of variables, always including those in a pre-specified set.
}
\details{
An error message is returned if \code{target} is smaller than the length of \code{active}.
}
\examples{
# generate linear regression data with 20 variables and 10 observations
res <- simData(prop=0.1, m=20, n=10, seed=42)

# choose target as twice the number of active variables
target <- 2*length(res$active)

# selection of at most target variables, including the active ones
targetOracle(res$X, res$Y, target, ncol(res$X), res$active)
}
\author{
Anna Vesely.
}
